requirements.txt
numpy==2.2.5
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
python-dotenv==1.0.0
requests==2.28.2